Create JdbiSqlExceptionMapper initial implementations. Move CWMS Database impl class to cwms subproject.
