+++ java/cwms-implementation/src/main/java/org/opendcs/cwms/CwmsOpenDcsDatabaseWrapper.java
@@ -1,14 +1,17 @@
-package decodes.cwms;
+package org.opendcs.cwms;
 
 import java.util.Map;
 
 import javax.sql.DataSource;
 
+import org.jdbi.v3.core.statement.SqlStatements;
 import org.opendcs.database.AbstractJdbiOpenDcsDatabaseWrapper;
 import org.opendcs.database.api.OpenDcsDataRuntimeException;
 import org.opendcs.database.dai.SiteReferenceMetaData;
+import org.opendcs.database.impl.cwms.jdbi.mapper.exception.CwmsExceptionMapper;
 import org.opendcs.settings.api.OpenDcsSettings;
 
+import decodes.cwms.CwmsLocationLevelDAO;
 import decodes.db.Database;
 import decodes.tsdb.TimeSeriesDb;
 
@@ -29,5 +32,7 @@ public class CwmsOpenDcsDatabaseWrapper extends AbstractJdbiOpenDcsDatabaseWrapp
         mapDao(SiteReferenceMetaData.class, wrapper);
         mapDao(CwmsLocationLevelDAO.class, wrapper);
         this.jdbi.define("numeric_date", false);
+        this.jdbi.getConfig(SqlStatements.class)
+                 .addExceptionHandler(new CwmsExceptionMapper());
     }
 }
