+++ java/cwms-implementation/src/main/java/org/opendcs/database/impl/cwms/jdbi/mapper/exception/CwmsExceptionMapper.java
@@ -1,4 +1,4 @@
-package org.opendcs.database.impl.opendcs.jdbi.mapper.exceptions;
+package org.opendcs.database.impl.cwms.jdbi.mapper.exception;
 
 import java.sql.SQLException;
 
@@ -6,13 +6,13 @@ import org.jdbi.v3.core.statement.SqlExceptionHandler;
 import org.jdbi.v3.core.statement.StatementContext;
 import org.opendcs.database.api.exceptions.data.RelatedDataConstraintException;
 
-public final class OpenDcsExceptionHandler implements SqlExceptionHandler
+public final class CwmsExceptionMapper implements SqlExceptionHandler
 {
 
     @Override
     public void handle(SQLException ex, StatementContext ctx)
     {
-        if (ex.getMessage().contains("violates foreign key"))
+        if (ex.getMessage().contains("integrity constraint"))
         {
             throw new RelatedDataConstraintException(ex);
         }
